pyproject.toml
[project]
name = "store-backend-service"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator==2.2.0",
    "fastapi==0.115.0",
    "google-cloud-firestore==2.19.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.4.0",
    "uvicorn[standard]==0.30.6",
]

[dependency-groups]
dev = [
    "pytest==8.3.3",
    "ruff==0.7.1",
]
